@@ -15,14 +15,17 @@
     "unpipe": "~1.0.0"
   },
   "devDependencies": {
-    "eslint": "3.18.0",
-    "eslint-config-standard": "7.1.0",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.2.0",
     "eslint-plugin-markdown": "1.0.0-beta.4",
+    "eslint-plugin-node": "4.2.2",
     "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "2.1.1",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "2.5.3",
     "readable-stream": "2.1.2",
+    "safe-buffer": "5.0.1",
     "supertest": "1.1.0"
   },
   "files": [
